Move doc processing helpers to different file
